+++ app/controllers/files.py
@@ -1,7 +1,3 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-from flask import Blueprint
 from flask import request, redirect
 from flask import render_template as template
 
@@ -11,22 +7,20 @@ from flask_security import permissions_required
 from app.helpers.helper_flask_view import HelperFlaskView
 from app.helpers.turbo_flash import turbo_flash as flash
 
-from ..models.files import File
-from ..handlers.files import FileHandler
+from app.models.files import File
 
-files_blueprint = Blueprint(
-    "files", __name__, url_prefix="/files", template_folder="templates"
-)
+from app.modules.files import show_file, download_file, all_files
 
 
 class FilesView(HelperFlaskView):
     def show(self, hash_value):
+
         file = File.load_by_attribute("hash", hash_value)
         thumbnail = request.args.get("thumbnail", False) == "True"
 
         self.validate_operation(hash_value, file)
 
-        return FileHandler().show(file, thumbnail=thumbnail)
+        return show_file(file, thumbnail)
 
     @route("/<id>/delete", methods=["POST"])
     def delete(self, id):
@@ -38,21 +32,12 @@ class FilesView(HelperFlaskView):
 
         return redirect(request.referrer)
 
-    # def show_profile_pic(self, user_id):
-    #     from app.models.users import User
-    #     file = User.load(user_id).profile_picture_file
-    #     if not file:
-    #         return FileHandler().show_new_user_placeholder()
-    #     if not file.can_view(current_user):
-    #         abort(403)
-    #     return FileHandler().show(file)
-
     def download(self, id):
         file = File.load(id)
         self.validate_operation(id, file)
 
-        return FileHandler().download(file)
+        return download_file(file)
 
     @permissions_required("manage-application")
     def index(self):
-        return template("files/index.html.j2", files=FileHandler().all_files)
+        return template("files/index.html.j2", files=all_files())
