--- app/modules/files/controllers/__init__.py
@@ -1,7 +1,0 @@
-from .files import FilesView
-
-__all__ = ["FilesView"]
-
-
-def register_all_controllers(application):
-    FilesView.register(application)
